requirements.txt
dash==2.14.2
gunicorn==23.0.0
gspread==5.12.4
mysql-connector-python==9.3.0
oauth2client==4.1.3
pandas==2.3.3
plotly==6.5.0
